[SERVICE-508] Start splitting contextChannels.inttest
