+++ test/demo/constants.ts
@@ -1,1 +1,15 @@
 export const appStartupTime = 30 * 1000;
+
+export const testManagerIdentity = {uuid: 'test-app', name: 'test-app'};
+
+export const testAppInDirectory1 = {
+    uuid: 'test-app-1',
+    name: 'test-app-1',
+    appId: '100'
+};
+
+export const testAppNotInDirectory = {
+    uuid: 'test-app-not-in-directory',
+    name: 'test-app-not-in-directory',
+    manifestUrl: 'http://localhost:3923/test/configs/test-app-not-in-directory.json'
+};
