+++ test/demo/contextChannels.inttest.ts
@@ -356,127 +356,3 @@ describe('When starting an app', () => {
         expect(payload[0]).toHaveProperty('identity', channelChangingWindow);
     }, appStartupTime * 2);
 });
-
-describe('When getting the list of desktop channels', () => {
-    test('All expected channels are returned', async () => {
-        const channels = await fdc3Remote.getDesktopChannels(testManagerIdentity);
-
-        const reducedChannels = channels.map(channel => ({id: channel.channel.id, type: channel.channel.type}));
-
-        expect(reducedChannels).toHaveLength(6);
-
-        expect(reducedChannels).toContainEqual({id: 'red', type: 'desktop'});
-        expect(reducedChannels).toContainEqual({id: 'orange', type: 'desktop'});
-        expect(reducedChannels).toContainEqual({id: 'yellow', type: 'desktop'});
-        expect(reducedChannels).toContainEqual({id: 'green', type: 'desktop'});
-        expect(reducedChannels).toContainEqual({id: 'blue', type: 'desktop'});
-        expect(reducedChannels).toContainEqual({id: 'purple', type: 'desktop'});
-
-        for (const channel of channels) {
-            expect(channel).toBeChannel({}, DesktopChannel);
-        }
-    });
-
-    test('Subsequent calls return the same channel instances', async () => {
-        const channels1 = await fdc3Remote.getDesktopChannels(testManagerIdentity);
-        const channels2 = await fdc3Remote.getDesktopChannels(testManagerIdentity);
-
-        expect(channels2).toHaveLength(channels1.length);
-
-        for (const channel of channels1) {
-            expect(channels2).toContain(channel);
-        }
-    });
-
-    test('When also querying a channel by ID, the same channel instance is returned', async () => {
-        const channels = await fdc3Remote.getDesktopChannels(testManagerIdentity);
-        const redChannel = await fdc3Remote.getChannelById(testManagerIdentity, 'red');
-
-        expect(channels).toContain(redChannel);
-    });
-});
-
-describe('When getting a channel by ID', () => {
-    test('When the ID provided is \'default\', the default channel is returned', async () => {
-        await expect(fdc3Remote.getChannelById(testManagerIdentity, 'default')).resolves.toBeChannel({id: 'default', type: 'default'}, DefaultChannel);
-    });
-
-    test('When the ID provided is \'green\', a desktop channel is returned', async () => {
-        await expect(fdc3Remote.getChannelById(testManagerIdentity, 'green')).resolves.toBeChannel({id: 'green', type: 'desktop'}, DesktopChannel);
-    });
-
-    test('Subsequent calls return the same channel instance', async () => {
-        const redChannel = await fdc3Remote.getChannelById(testManagerIdentity, 'red');
-        await expect(fdc3Remote.getChannelById(testManagerIdentity, 'red')).resolves.toBe(redChannel);
-    });
-
-    test('If the channel does not exist, an FDC3Error is thrown', async () => {
-        const getChannelByIdPromise = fdc3Remote.getChannelById(testManagerIdentity, 'non-existent-channel');
-
-        await expect(getChannelByIdPromise).toThrowFDC3Error(
-            ChannelError.ChannelDoesNotExist,
-            'No channel with channelId: non-existent-channel'
-        );
-    });
-});
-
-describe('When querying the current channel', () => {
-    test('If no identity is provided, the channel of the current window is returned', async () => {
-        const [redWindow] = await setupWindows('red');
-
-        await expect(fdc3Remote.getCurrentChannel(redWindow)).resolves.toBeChannel({id: 'red', type: 'desktop'}, DesktopChannel);
-    });
-
-    test('If querying the channel of a different window, the correct result is returned', async () => {
-        const [purpleWindow, greenWindow] = await setupWindows('purple', 'green');
-
-        await expect(fdc3Remote.getCurrentChannel(purpleWindow, greenWindow)).resolves.toBeChannel({id: 'green', type: 'desktop'}, DesktopChannel);
-    });
-
-    test('Subsequent calls return the same channel instance', async () => {
-        const [blueWindow] = await setupWindows('blue');
-
-        const blueChannel = await fdc3Remote.getCurrentChannel(blueWindow);
-        await expect(fdc3Remote.getCurrentChannel(blueWindow)).resolves.toBe(blueChannel);
-    });
-
-    test('If an invalid identity is provided, a TypeError is thrown', async () => {
-        const invalidIdentity: Identity = {irrelevantProperty: 'irrelevantValue'} as unknown as Identity;
-
-        await expect(fdc3Remote.getCurrentChannel(testManagerIdentity, invalidIdentity)).rejects
-            .toThrowError(new TypeError(`${JSON.stringify(invalidIdentity)} is not a valid Identity`));
-    });
-
-    test('If an identity for a window that does not exist is provided, an FDC3 error is thrown', async () => {
-        const nonExistentWindowIdentity: Identity = {uuid: 'does-not-exist', name: 'does-not-exist'};
-
-        await expect(fdc3Remote.getCurrentChannel(testManagerIdentity, nonExistentWindowIdentity)).toThrowFDC3Error(
-            IdentityError.WindowWithIdentityNotFound,
-            `No connection to FDC3 service found from window with identity: ${JSON.stringify(nonExistentWindowIdentity)}`
-        );
-    });
-
-    describe('When a non-FDC3 app is running', () => {
-        const testAppNotFdc3 = {
-            uuid: 'test-app-not-fdc3',
-            name: 'test-app-not-fdc3',
-            manifestUrl: 'http://localhost:3923/test/configs/test-app-not-fdc3.json'
-        };
-
-        beforeEach(async () => {
-            await fin.Application.startFromManifest(testAppNotFdc3.manifestUrl);
-        });
-
-        afterEach(async () => {
-            await fin.Application.wrapSync(testAppNotFdc3).quit(true);
-        });
-
-        test('If the non-FDC3 window identity is provided, an FDC3 error is thrown', async () => {
-            await expect(fdc3Remote.getCurrentChannel(testManagerIdentity, testAppNotFdc3)).toThrowFDC3Error(
-                IdentityError.WindowWithIdentityNotFound,
-                `No connection to FDC3 service found from window with identity: \
-${JSON.stringify({uuid: testAppNotFdc3.uuid, name: testAppNotFdc3.name})}`
-            );
-        });
-    });
-});
